Remove collections dependency and es6-set instead
Since [collections module][1] modifies native objects, it's only good
to be used as a top dependency of an application and not modules that
it's using.

The problem is that this will conflict if native objects has been
modified by another module, let's say something like [sugar][2] or just
another version of collections. When you have two of these being used in
your app, you might be in trouble, where some code depends on a
behaviour of a method implemented by one module but actually the method
had been silently overriden by another module having a different
behaviour.

[1]: http://collectionsjs.com
[2]: http://sugarjs.com