Implements buffer decoding in HTTP.read()
When HTTP responses arrive as buffers there is no safe way for the user
of HTTP.read to determine the buffer encoding.

The read method however can use the `Content-Type` header to determine
content encoding and convert the respose to string safely. This change
makes read behave according to the API docs, which state that read()
returns a string.

Introduces dependencies on iconv-lite and strip-bom.