Fix incorrect version check
As it stands, the version check will ensure that the major version is
non-negative AND that the minor version is at least 10.  Current versions,
with major version 4 or 5 but single-digit minor versions, are not matched
by this.  Instead we can check whether the major version is at least 1 OR
the minor version is at least 10, assuming in the latter case that the major
version cannot be less than 0 so we don't have to check for that.